add rtt check

+++ nack-test/receiver/Transport/test_tp.h
@@ -18,7 +18,7 @@ namespace transportdemo {
     uint32_t length;
     uint16_t padding;
 
-    uint32_t get_type() const { return ntohs(type); }
+    uint16_t get_type() const { return ntohs(type); }
     uint32_t get_length() const { return ntohl(length); }
   };
 
@@ -27,7 +27,7 @@ namespace transportdemo {
     uint16_t sequence;
     uint32_t timestamp;
 
-    uint32_t get_type() const { return ntohs(type); }
+    uint16_t get_type() const { return ntohs(type); }
     uint16_t get_sequence() const { return ntohs(sequence); }
     uint32_t get_timestamp() const { return ntohl(timestamp); }
   };
